requirements.txt
SQLAlchemy==1.4.41
flask-bootstrap4==4.0.2
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.1.0
gevent==22.10.2
itsdangerous==2.0.1
passlib==1.7.4
psycopg2==2.9.1
wtforms==2.3.3
wtforms-json==0.3.3
python-dotenv==0.17.1
flask-migrate==3.1.0
boto3==1.20.24
werkzeug==2.0.2
Jinja2==3.0.3
PyYAML==5.4
sentry-sdk[flask, sqlalchemy]==1.17.0